Closes #19829: Move object types REST API endpoint to core app
