+++ netbox/core/api/serializers.py
@@ -1,4 +1,5 @@
 from .serializers_.change_logging import *
 from .serializers_.data import *
 from .serializers_.jobs import *
+from .serializers_.object_types import *
 from .serializers_.tasks import *
