+++ netbox/core/api/serializers_/object_types.py
