+++ netbox/core/tests/test_api.py
@@ -7,6 +7,7 @@ from django.utils import timezone
 from rq.job import Job as RQ_Job, JobStatus
 from rq.registry import FailedJobRegistry, StartedJobRegistry
 
+from rest_framework import status
 from users.models import Token, User
 from utilities.testing import APITestCase, APIViewTestCases, TestCase
 from utilities.testing.utils import disable_logging
@@ -99,6 +100,22 @@ class DataFileTest(
             ),
         )
         DataFile.objects.bulk_create(data_files)
+
+
+class ObjectTypeTest(APITestCase):
+
+    def test_list_objects(self):
+        object_type_count = ObjectType.objects.count()
+
+        response = self.client.get(reverse('extras-api:objecttype-list'), **self.header)
+        self.assertHttpStatus(response, status.HTTP_200_OK)
+        self.assertEqual(response.data['count'], object_type_count)
+
+    def test_get_object(self):
+        object_type = ObjectType.objects.first()
+
+        url = reverse('extras-api:objecttype-detail', kwargs={'pk': object_type.pk})
+        self.assertHttpStatus(self.client.get(url, **self.header), status.HTTP_200_OK)
 
 
 class BackgroundTaskTestCase(TestCase):
